requirements.txt
asttokens==3.0.0
backcall==0.2.0
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
executing==2.2.0
importlib-metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
jupyter-client==8.6.3
jupyter-core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.24.4
packaging==24.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.6
prompt-toolkit==3.0.50
psutil==6.1.1
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2025.1
wcwidth==0.2.13
zipp==3.20.2
